docker/flask/requirements.txt
flask
flask-socketio
psycopg2-binary
flask_sqlalchemy
grpcio
grpcio-tools